feat(restructureFolder) - remove commented out files - restructure folder [Delivers: incomplete]
