+++ client/src/components/account/smart/Login.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import { PropTypes } from 'prop-types';
-import HomePage from '../../home/presentational/HomePage.jsx';
+import HomePage from '../../home/smart/HomePage.jsx';
 import { LoginAction } from '../../../actions/authActions.js';
 import { addFlashMessage } from '../../../actions/flashMessages.js';
