+++ client/src/components/home/smart/HomePage.jsx
